Move protocol implementation into new package
