+++ com.microsoft.java.debug.core/src/main/java/com/microsoft/java/debug/core/UsageDataSession.java
@@ -21,9 +21,9 @@ import java.util.logging.Logger;
 
 import com.google.gson.JsonElement;
 import com.microsoft.java.debug.core.adapter.AdapterUtils;
-import com.microsoft.java.debug.core.adapter.JsonUtils;
-import com.microsoft.java.debug.core.adapter.Messages.Request;
-import com.microsoft.java.debug.core.adapter.Messages.Response;
+import com.microsoft.java.debug.core.protocol.JsonUtils;
+import com.microsoft.java.debug.core.protocol.Messages.Request;
+import com.microsoft.java.debug.core.protocol.Messages.Response;
 import com.sun.jdi.event.Event;
 
 public class UsageDataSession {
