+++ com.microsoft.java.debug.core/src/main/java/com/microsoft/java/debug/core/adapter/handler/ThreadsRequestHandler.java
@@ -18,21 +18,21 @@ import java.util.List;
 import com.microsoft.java.debug.core.DebugUtility;
 import com.microsoft.java.debug.core.adapter.AdapterUtils;
 import com.microsoft.java.debug.core.adapter.ErrorCode;
-import com.microsoft.java.debug.core.adapter.Events;
 import com.microsoft.java.debug.core.adapter.IDebugAdapterContext;
 import com.microsoft.java.debug.core.adapter.IDebugRequestHandler;
-import com.microsoft.java.debug.core.adapter.Messages.Response;
-import com.microsoft.java.debug.core.adapter.Requests;
-import com.microsoft.java.debug.core.adapter.Requests.Arguments;
-import com.microsoft.java.debug.core.adapter.Requests.Command;
-import com.microsoft.java.debug.core.adapter.Requests.ContinueArguments;
-import com.microsoft.java.debug.core.adapter.Requests.NextArguments;
-import com.microsoft.java.debug.core.adapter.Requests.PauseArguments;
-import com.microsoft.java.debug.core.adapter.Requests.StepInArguments;
-import com.microsoft.java.debug.core.adapter.Requests.StepOutArguments;
-import com.microsoft.java.debug.core.adapter.Requests.ThreadsArguments;
-import com.microsoft.java.debug.core.adapter.Responses;
-import com.microsoft.java.debug.core.adapter.Types;
+import com.microsoft.java.debug.core.protocol.Events;
+import com.microsoft.java.debug.core.protocol.Messages.Response;
+import com.microsoft.java.debug.core.protocol.Requests;
+import com.microsoft.java.debug.core.protocol.Requests.Arguments;
+import com.microsoft.java.debug.core.protocol.Requests.Command;
+import com.microsoft.java.debug.core.protocol.Requests.ContinueArguments;
+import com.microsoft.java.debug.core.protocol.Requests.NextArguments;
+import com.microsoft.java.debug.core.protocol.Requests.PauseArguments;
+import com.microsoft.java.debug.core.protocol.Requests.StepInArguments;
+import com.microsoft.java.debug.core.protocol.Requests.StepOutArguments;
+import com.microsoft.java.debug.core.protocol.Requests.ThreadsArguments;
+import com.microsoft.java.debug.core.protocol.Responses;
+import com.microsoft.java.debug.core.protocol.Types;
 import com.sun.jdi.ObjectCollectedException;
 import com.sun.jdi.ThreadReference;
 import com.sun.jdi.VMDisconnectedException;
